// IntraFAD/src/app/models/docente.model.ts
export interface Docente {
    _id: String;
    img: String;
    nombre: String;
    apellido_p: String;
    apellido_m: String;
    direccion: String;
    correo_personal: String;
    correo_institucional: String;
    telefono: String;
    rfc: String;
    doc_rfc: String;
    curp: String;
    doc_curp: String;
    rol: String;
    no_empleado: String;
    ldg: Boolean;
    ldi: Boolean;
    arq: Boolean;
    apou: Boolean;
}